add class to handle the processTerminated event
This class is guaranteed invoked after the pitest task has finished
so it's guaranteed that results are present (if any have been written).

Fixes an issue where sometimes the annotator was restarted before the pitest run was finished.

This also only restarts parsers for files that have been affected by the last pitest run.

Signed-off-by: Lennart Heimbs <lennart@heimbs.me>