revamp the entire plugin checking part.
This completely dumps any parsing of files in favor of a gradle task
registered by the compagnion plugin which existence signals the
presence of the compagnion. In turn the compagnion task check for the
solidsoft and driodsondriods plugins and prints then in json form
in the gradle task. If we then parse the output of that task
we get the presence of the pitest plugins.

The gradle task is also able to get the available android build types.
Sadly i still haven't found a way to get the currently set build type.
I think that information mainly lives in the intellij android plugin...

Signed-off-by: Lennart Heimbs <lennart@heimbs.me>